+++ src/Facades/FilamentOdometerEasy.php
@@ -6,8 +6,8 @@ use Illuminate\Support\Facades\Facade;
 use Illuminate\Support\HtmlString;
 
 /**
- * @method static HtmlString render(mixed $value, string|array|null $format = null, ?string $class = null, ?string $driver = null)
- * @method static HtmlString renderNumberFlow(mixed $value, ?array $format = null, ?string $class = null)
+ * @method static HtmlString render(mixed $value, string|array|null $format = null, ?string $class = null, ?string $driver = null, ?int $duration = null)
+ * @method static HtmlString renderNumberFlow(mixed $value, ?array $format = null, ?string $class = null, ?int $delay = null, ?int $duration = null)
  * @method static HtmlString renderOdometer(mixed $value, ?string $format = null, ?string $class = null)
  *
  * @see \Gsferro\FilamentOdometerEasy\FilamentOdometerEasy